requirements.txt
django
psycopg[binary]
asgiref==3.8.1
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
dj-database-url